src/requirements.txt
Flask
SQLAlchemy
Flask-SQLAlchemy
psycopg2-binary
Flask-Cors
python-dotenv
PyJWT==2.8.0